Rename execute_action_capability -> execute_path_action_capability
