+++ move_group/CMakeLists.txt
@@ -47,7 +47,7 @@ add_executable(list_move_group_capabilities src/list_capabilities.cpp)
 
 add_library(moveit_move_group_default_capabilities
   src/default_capabilities/move_action_capability.cpp
-  src/default_capabilities/execute_action_capability.cpp
+  src/default_capabilities/execute_path_action_capability.cpp
   src/default_capabilities/plan_service_capability.cpp
   src/default_capabilities/execute_service_capability.cpp
   src/default_capabilities/query_planners_service_capability.cpp
